add ArkBootRunner test cases
