+++ sofa-ark-parent/support/ark-ide-starter/src/test/resources/META-INF/sofa-ark-test/sofa-ark-test.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans xmlns="http://www.springframework.org/schema/beans"
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
+
+    <bean id="sampleService" class="com.alipay.sofa.ark.ide.springboot.impl.SampleServiceImpl"/>
+
+</beans>
